新增更新bgi功能

+++ BetterGI/uploadBgi.go
@@ -4,6 +4,7 @@ import (
 	"auto-bgi/autoLog"
 	"auto-bgi/bgiStatus"
 	"auto-bgi/config"
+	"auto-bgi/control"
 	"auto-bgi/tools"
 	"github.com/gin-gonic/gin"
 	abgiCopy "github.com/otiai10/copy"
@@ -16,6 +17,8 @@ import (
 
 // 上传bgi压缩包
 func UploadBgi(c *gin.Context) {
+
+	control.CloseSoftware()
 
 	// 单文件上传
 	file, err := c.FormFile("file")
@@ -33,6 +36,11 @@ func UploadBgi(c *gin.Context) {
 	time.Sleep(1 * time.Second)
 
 	UpdateBgi()
+
+	//更新仓库
+	go func() {
+		bgiStatus.GitPull()
+	}()
 
 	c.JSON(http.StatusOK, gin.H{
 		"message":  "更新成功",
@@ -54,6 +62,14 @@ func UpdateBgi() {
 	}
 	autoLog.Sugar.Infof("备份user文件")
 
+	//备份log
+	err2 := abgiCopy.Copy(config.Cfg.BetterGIAddress+"\\log\\", "backups\\log\\")
+	if err2 != nil {
+		autoLog.Sugar.Errorf("备份log失败: %v", err2)
+		return
+	}
+	autoLog.Sugar.Infof("备份log")
+
 	//2、删除bgi文件夹
 	err := os.RemoveAll(config.Cfg.BetterGIAddress)
 	if err != nil {
@@ -65,11 +81,6 @@ func UpdateBgi() {
 	//3、解压bgi压缩包
 	base := filepath.Base(config.Cfg.BetterGIAddress)
 	path := strings.ReplaceAll(config.Cfg.BetterGIAddress, base, "")
-	//err = tools.Unzip("./uploads/BetterGI.zip", path)
-	//if err != nil {
-	//	autoLog.Sugar.Errorf("解压bgi压缩包失败: %v", err)
-	//	return
-	//}
 	err4 = tools.Un7z("./uploads/BetterGI.zip", path)
 	if err4 != nil {
 		autoLog.Sugar.Errorf("解压bgi压缩包失败: %v", err)
@@ -79,7 +90,7 @@ func UpdateBgi() {
 	autoLog.Sugar.Infof("解压bgi压缩包")
 
 	//4、删除user文件夹
-	err2 := os.RemoveAll(config.Cfg.BetterGIAddress + "\\User\\")
+	err2 = os.RemoveAll(config.Cfg.BetterGIAddress + "\\User\\")
 	if err2 != nil {
 		autoLog.Sugar.Errorf("删除user文件夹失败: %v", err2)
 		return
@@ -108,9 +119,24 @@ func UpdateBgi() {
 		autoLog.Sugar.Errorf("删除user压缩包失败: %v", err5)
 	}
 
+	//8、还原log
+	err5 = abgiCopy.Copy("backups\\log\\", config.Cfg.BetterGIAddress+"\\log\\")
+	if err5 != nil {
+		autoLog.Sugar.Errorf("还原log失败: %v", err5)
+	}
+	autoLog.Sugar.Infof("还原log")
+
 	err5 = os.Remove("./uploads/BetterGI.zip")
 	if err5 != nil {
 		autoLog.Sugar.Errorf("删除BetterGI压缩包失败: %v", err5)
 	}
+	autoLog.Sugar.Infof("删除BetterGI压缩包")
+
+	//删除log
+	err5 = os.RemoveAll("backups\\log\\")
+	if err5 != nil {
+		autoLog.Sugar.Errorf("删除log失败: %v", err5)
+	}
+	autoLog.Sugar.Infof("删除log")
 
 }
